attendance page edit 분리
