+++ src/pages/AttendancePage/CourseAttendanceCard/style.jsx
@@ -1,8 +1,9 @@
 import styled from 'styled-components';
+
 import { StyledSidePadding } from '../../../utility/COMMON_STYLE';
 
 export const StyledContainer = styled(StyledSidePadding)`
-  background-color: rgb(245, 245, 245);
+  background-color: $ ${BASE_COLOR};
   display: grid;
   grid-template-columns: 130px auto;
   padding-bottom: 40px;
@@ -19,7 +20,6 @@ export const StyledBox = styled.div`
   height: 130px;
   display: grid;
   place-items: center;
-  box-shadow: rgba(0, 0, 0, 0.187) 0px 4px 5px 1px;
   cursor: pointer;
 `;
 
@@ -30,7 +30,6 @@ export const StyledEmoji = styled.div`
 
 export const StyledAttendanceBox = styled.div`
   background-color: white;
-  box-shadow: inset rgba(0, 0, 0, 0.187) 0px 3px 1.5px;
   margin-left: 60px;
   height: 60px;
   border-radius: 30px;
